Bean madness

+++ Carportal/src/main/java/adessoassessment/carportal/restservice/FeatureRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 @Repository
-public interface FeatureRepository extends JpaRepository<Features, Long>{
+public interface FeatureRepository extends JpaRepository<Features, Integer>{
 }
